add ping WAN function.

+++ settings.php
@@ -5,25 +5,36 @@
     $type = $data["type"];
 
     if ($categories == "network") {
-        $networkFile = fopen("/etc/network/interfaces", "w+");
-        $fileString = "";
-        if ($type == "dhcp") {
-            $fileString = "auto lo \niface lo inet loopback\n\n#dhcp \nauto eth0\niface eth0 inet dhcp\n";
-        } else {
-            $fileString = $fileString ."auto lo \niface lo inet loopback\n\n";
-            $fileString = $fileString ."auto eth0\niface eth0 inet static\n";
-            $fileString = $fileString ."    address ".$data["ip"]."\n";
-            $fileString = $fileString ."    netmask ".$data["netmask"]."\n";
-            $fileString = $fileString ."    broadcast ".$data["broadcast"]."\n";
-            $fileString = $fileString ."    gateway ".$data["gateway"]."\n";
+        if ($type == "dhcp" || $type == "staticIP") {
+            $networkFile = fopen("/etc/network/interfaces", "w+");
+            $fileString = "";
+            if ($type == "dhcp") {
+                $fileString = "auto lo \niface lo inet loopback\n\n#dhcp \nauto eth0\niface eth0 inet dhcp\n";
+            } elseif ($type == "staticIP"){
+                $fileString = $fileString ."auto lo \niface lo inet loopback\n\n";
+                $fileString = $fileString ."auto eth0\niface eth0 inet static\n";
+                $fileString = $fileString ."    address ".$data["ip"]."\n";
+                $fileString = $fileString ."    netmask ".$data["netmask"]."\n";
+                $fileString = $fileString ."    broadcast ".$data["broadcast"]."\n";
+                $fileString = $fileString ."    gateway ".$data["gateway"]."\n";
+            } 
+            fwrite($networkFile, $fileString);
+            fclose($networkFile);
+            
+            echo '{"status": "ok"}';
+            
+            shell_exec('sync');
+            shell_exec('reboot');
         }
-        fwrite($networkFile, $fileString);
-        fclose($networkFile);
-        
-        echo '{"status": "ok"}';
-        
-    	shell_exec('sync');
-    	shell_exec('reboot');
+
+        if ($type == "ping") {
+            $IPOrDNS = $data["IPOrDNS"];
+            $result = exec("ping -c 1 '".$IPOrDNS."' 2>&1 | grep '0% packet loss,'");
+            if ($result != null)
+                echo '{"status": "ok"}';
+            else
+                echo '{"status": "error"}';
+        }
     }
 
     if ($categories == "dateAndTime" && $type == "dateAndTime") {
